Update test for both replication and galera. Refactor MariaDB tests
