+++ controller/mariadb_controller_replication_test.go
@@ -1,0 +1,300 @@
+package controller
+
+import (
+	"time"
+
+	mariadbv1alpha1 "github.com/mariadb-operator/mariadb-operator/api/v1alpha1"
+	"github.com/mariadb-operator/mariadb-operator/pkg/statefulset"
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+	corev1 "k8s.io/api/core/v1"
+	policyv1 "k8s.io/api/policy/v1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/apimachinery/pkg/api/resource"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/utils/ptr"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+var _ = Describe("MariaDB replication", func() {
+	Context("When creating a MariaDB with replication", func() {
+		It("Should reconcile", func() {
+			mdb := mariadbv1alpha1.MariaDB{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "mariadb-repl",
+					Namespace: testNamespace,
+				},
+				Spec: mariadbv1alpha1.MariaDBSpec{
+					Username: &testUser,
+					PasswordSecretKeyRef: &mariadbv1alpha1.GeneratedSecretKeyRef{
+						SecretKeySelector: corev1.SecretKeySelector{
+							LocalObjectReference: corev1.LocalObjectReference{
+								Name: testPwdKey.Name,
+							},
+							Key: testPwdSecretKey,
+						},
+					},
+					Database: &testDatabase,
+					MyCnf: func() *string {
+						cfg := `[mariadb]
+						bind-address=*
+						default_storage_engine=InnoDB
+						binlog_format=row
+						innodb_autoinc_lock_mode=2
+						max_allowed_packet=256M`
+						return &cfg
+					}(),
+					Replication: &mariadbv1alpha1.Replication{
+						ReplicationSpec: mariadbv1alpha1.ReplicationSpec{
+							Primary: &mariadbv1alpha1.PrimaryReplication{
+								PodIndex:          func() *int { i := 0; return &i }(),
+								AutomaticFailover: func() *bool { f := true; return &f }(),
+							},
+							Replica: &mariadbv1alpha1.ReplicaReplication{
+								WaitPoint: func() *mariadbv1alpha1.WaitPoint { w := mariadbv1alpha1.WaitPointAfterSync; return &w }(),
+								Gtid:      func() *mariadbv1alpha1.Gtid { g := mariadbv1alpha1.GtidCurrentPos; return &g }(),
+							},
+							SyncBinlog: func() *bool { s := true; return &s }(),
+						},
+						Enabled: true,
+					},
+					Replicas: 3,
+					Storage: mariadbv1alpha1.Storage{
+						Size:                ptr.To(resource.MustParse("300Mi")),
+						StorageClassName:    "standard-resize",
+						ResizeInUseVolumes:  ptr.To(true),
+						WaitForVolumeResize: ptr.To(true),
+					},
+					Service: &mariadbv1alpha1.ServiceTemplate{
+						Type: corev1.ServiceTypeLoadBalancer,
+						Metadata: &mariadbv1alpha1.Metadata{
+							Annotations: map[string]string{
+								"metallb.universe.tf/loadBalancerIPs": testCidrPrefix + ".0.120",
+							},
+						},
+					},
+					Connection: &mariadbv1alpha1.ConnectionTemplate{
+						SecretName: func() *string {
+							s := "mdb-repl-conn"
+							return &s
+						}(),
+						SecretTemplate: &mariadbv1alpha1.SecretTemplate{
+							Key: &testConnSecretKey,
+						},
+					},
+					PrimaryService: &mariadbv1alpha1.ServiceTemplate{
+						Type: corev1.ServiceTypeLoadBalancer,
+						Metadata: &mariadbv1alpha1.Metadata{
+							Annotations: map[string]string{
+								"metallb.universe.tf/loadBalancerIPs": testCidrPrefix + ".0.130",
+							},
+						},
+					},
+					PrimaryConnection: &mariadbv1alpha1.ConnectionTemplate{
+						SecretName: func() *string {
+							s := "mdb-repl-conn-primary"
+							return &s
+						}(),
+						SecretTemplate: &mariadbv1alpha1.SecretTemplate{
+							Key: &testConnSecretKey,
+						},
+					},
+					SecondaryService: &mariadbv1alpha1.ServiceTemplate{
+						Type: corev1.ServiceTypeLoadBalancer,
+						Metadata: &mariadbv1alpha1.Metadata{
+							Annotations: map[string]string{
+								"metallb.universe.tf/loadBalancerIPs": testCidrPrefix + ".0.131",
+							},
+						},
+					},
+					SecondaryConnection: &mariadbv1alpha1.ConnectionTemplate{
+						SecretName: func() *string {
+							s := "mdb-repl-conn-secondary"
+							return &s
+						}(),
+						SecretTemplate: &mariadbv1alpha1.SecretTemplate{
+							Key: &testConnSecretKey,
+						},
+					},
+					UpdateStrategy: mariadbv1alpha1.UpdateStrategy{
+						Type: mariadbv1alpha1.ReplicasFirstPrimaryLast,
+					},
+				},
+			}
+
+			By("Creating MariaDB with replication")
+			Expect(k8sClient.Create(testCtx, &mdb)).To(Succeed())
+			DeferCleanup(func() {
+				deleteMariaDB(&mdb)
+			})
+
+			By("Expecting MariaDB to be ready eventually")
+			Eventually(func() bool {
+				if err := k8sClient.Get(testCtx, client.ObjectKeyFromObject(&mdb), &mdb); err != nil {
+					return false
+				}
+				return mdb.IsReady()
+			}, testHighTimeout, testInterval).Should(BeTrue())
+
+			By("Expecting to create a Service")
+			var svc corev1.Service
+			Expect(k8sClient.Get(testCtx, client.ObjectKeyFromObject(&mdb), &svc)).To(Succeed())
+
+			By("Expecting to create a primary Service")
+			Expect(k8sClient.Get(testCtx, mdb.PrimaryServiceKey(), &svc)).To(Succeed())
+			Expect(svc.Spec.Selector["statefulset.kubernetes.io/pod-name"]).To(Equal(statefulset.PodName(mdb.ObjectMeta, 0)))
+
+			By("Expecting to create a secondary Service")
+			Expect(k8sClient.Get(testCtx, mdb.SecondaryServiceKey(), &svc)).To(Succeed())
+
+			By("Expecting Connection to be ready eventually")
+			Eventually(func() bool {
+				var conn mariadbv1alpha1.Connection
+				if err := k8sClient.Get(testCtx, client.ObjectKeyFromObject(&mdb), &conn); err != nil {
+					return false
+				}
+				return conn.IsReady()
+			}, testTimeout, testInterval).Should(BeTrue())
+
+			By("Expecting primary Connection to be ready eventually")
+			Eventually(func() bool {
+				var conn mariadbv1alpha1.Connection
+				if err := k8sClient.Get(testCtx, mdb.PrimaryConnectioneKey(), &conn); err != nil {
+					return false
+				}
+				return conn.IsReady()
+			}, testTimeout, testInterval).Should(BeTrue())
+
+			By("Expecting secondary Connection to be ready eventually")
+			Eventually(func() bool {
+				var conn mariadbv1alpha1.Connection
+				if err := k8sClient.Get(testCtx, mdb.SecondaryConnectioneKey(), &conn); err != nil {
+					return false
+				}
+				return conn.IsReady()
+			}, testTimeout, testInterval).Should(BeTrue())
+
+			By("Expecting to create secondary Endpoints")
+			var endpoints corev1.Endpoints
+			Expect(k8sClient.Get(testCtx, mdb.SecondaryServiceKey(), &endpoints)).To(Succeed())
+			Expect(endpoints.Subsets).To(HaveLen(1))
+			Expect(endpoints.Subsets[0].Addresses).To(HaveLen(int(mdb.Spec.Replicas) - 1))
+
+			By("Expecting to create a PodDisruptionBudget")
+			var pdb policyv1.PodDisruptionBudget
+			Expect(k8sClient.Get(testCtx, client.ObjectKeyFromObject(&mdb), &pdb)).To(Succeed())
+
+			By("Failing and switching over primary")
+			testReplicationFailover(&mdb)
+
+			By("Updating MariaDB")
+			testMariadbUpdate(&mdb, "mariadb:lts")
+
+			By("Resizing MariaDB PVCs")
+			testMariadbVolumeResize(&mdb, "400Mi")
+
+			mxsKey := types.NamespacedName{
+				Name:      "maxscale-repl",
+				Namespace: testNamespace,
+			}
+			By("Using MariaDB with MaxScale")
+			testMariadbMaxscale(&mdb, mxsKey)
+		})
+	})
+})
+
+func testReplicationFailover(mdb *mariadbv1alpha1.MariaDB) {
+	key := client.ObjectKeyFromObject(mdb)
+
+	By("Expecting MariaDB primary to be set")
+	Eventually(func() bool {
+		return mdb.Status.CurrentPrimary != nil
+	}, testTimeout, testInterval).Should(BeTrue())
+
+	currentPrimary := *mdb.Status.CurrentPrimary
+	By("Tearing down primary Pod consistently")
+	Consistently(func() bool {
+		primaryPodKey := types.NamespacedName{
+			Name:      currentPrimary,
+			Namespace: mdb.Namespace,
+		}
+		var primaryPod corev1.Pod
+		if err := k8sClient.Get(testCtx, primaryPodKey, &primaryPod); err != nil {
+			return apierrors.IsNotFound(err)
+		}
+		return k8sClient.Delete(testCtx, &primaryPod) == nil
+	}, 10*time.Second, testInterval).Should(BeTrue())
+
+	By("Expecting MariaDB to be ready eventually")
+	Eventually(func() bool {
+		if err := k8sClient.Get(testCtx, key, mdb); err != nil {
+			return false
+		}
+		return mdb.IsReady()
+	}, testHighTimeout, testInterval).Should(BeTrue())
+
+	By("Expecting MariaDB to eventually change primary")
+	Eventually(func() bool {
+		if err := k8sClient.Get(testCtx, key, mdb); err != nil {
+			return false
+		}
+		if !mdb.IsReady() || mdb.Status.CurrentPrimary == nil {
+			return false
+		}
+		return *mdb.Status.CurrentPrimary != currentPrimary
+	}, testHighTimeout, testInterval).Should(BeTrue())
+
+	By("Expecting Connection to be ready eventually")
+	Eventually(func() bool {
+		var conn mariadbv1alpha1.Connection
+		if err := k8sClient.Get(testCtx, key, &conn); err != nil {
+			return false
+		}
+		return conn.IsReady()
+	}, testTimeout, testInterval).Should(BeTrue())
+
+	By("Expecting primary Connection to be ready eventually")
+	Eventually(func() bool {
+		var conn mariadbv1alpha1.Connection
+		if err := k8sClient.Get(testCtx, mdb.PrimaryConnectioneKey(), &conn); err != nil {
+			return false
+		}
+		return conn.IsReady()
+	}, testTimeout, testInterval).Should(BeTrue())
+
+	By("Expecting MariaDB to eventually update primary")
+	var podIndex int
+	for i := 0; i < int(mdb.Spec.Replicas); i++ {
+		if i != *mdb.Status.CurrentPrimaryPodIndex {
+			podIndex = i
+			break
+		}
+	}
+	Eventually(func(g Gomega) bool {
+		g.Expect(k8sClient.Get(testCtx, key, mdb)).To(Succeed())
+		mdb.Replication().Primary.PodIndex = &podIndex
+		g.Expect(k8sClient.Update(testCtx, mdb)).To(Succeed())
+		return true
+	}, testTimeout, testInterval).Should(BeTrue())
+
+	By("Expecting MariaDB to eventually change primary")
+	Eventually(func() bool {
+		if err := k8sClient.Get(testCtx, key, mdb); err != nil {
+			return false
+		}
+		if !mdb.IsReady() || mdb.Status.CurrentPrimaryPodIndex == nil {
+			return false
+		}
+		return *mdb.Status.CurrentPrimaryPodIndex == podIndex
+	}, testTimeout, testInterval).Should(BeTrue())
+
+	By("Expecting primary Service to eventually change primary")
+	Eventually(func() bool {
+		var svc corev1.Service
+		if err := k8sClient.Get(testCtx, mdb.PrimaryServiceKey(), &svc); err != nil {
+			return false
+		}
+		return svc.Spec.Selector["statefulset.kubernetes.io/pod-name"] == statefulset.PodName(mdb.ObjectMeta, podIndex)
+	}, testTimeout, testInterval).Should(BeTrue())
+}
